integrated the recyclerview into the app. Added the ContentFragment, also loads the needed data. Next step: charts

+++ app/src/main/java/com/example/bachelorthesis/persistence/DAOs/PatientDAO.java
@@ -31,7 +31,7 @@ public interface PatientDAO {
 
     @Transaction
     @Query("Select * from patient where id = :id")
-    PatientWithData findPatientWithData(int id);
+    PatientWithData findPatientWithData(long id);
 
     @Query("Delete from patient")
     void nukePatients();
